move pipeFileContents into test-only file
