+++ test/slpStream.spec.ts
@@ -1,6 +1,7 @@
 import * as sinon from "sinon";
 
-import { RxSlpStream, pipeFileContents } from "../src";
+import { RxSlpStream } from "../src";
+import { pipeFileContents } from "./pipeFileContents";
 import { Subscription } from "rxjs";
 import { SlpStreamMode } from "@slippi/slippi-js";
 
@@ -20,7 +21,7 @@ describe("SlpStream", () => {
       const gameStartSpy = sinon.spy();
       const gameEndSpy = sinon.spy();
 
-      const slpStream = new RxSlpStream({ mode: SlpStreamMode.MANUAL });
+      const slpStream = new RxSlpStream({ suppressErrors: false, mode: SlpStreamMode.MANUAL });
       const unsubGameStart = slpStream.gameStart$.subscribe(gameStartSpy);
       const unsubGameEnd = slpStream.gameEnd$.subscribe(gameEndSpy);
       subscriptions.push(unsubGameStart, unsubGameEnd);
