+++ test/winner.spec.ts
@@ -1,4 +1,5 @@
-import { pipeFileContents, SlpRealTime, RxSlpStream } from "../src";
+import { SlpRealTime, RxSlpStream } from "../src";
+import { pipeFileContents } from "./pipeFileContents";
 import { SlpStreamMode } from "@slippi/slippi-js";
 import { Subscription } from "rxjs";
